[CRITICAL] prevent unauthorized access
Prior to this commit, it is possible to access the entire userbot through an inline bot, whether or not it is controlled by the user. This prevents the userbot from reacting to commands sent via inline bot, as otherwise a devastating exploit is possible with only a tiny amount of social engineering.